fix: reduce CRM backfill duplication
