+++ apps/web/src/app/api/cron/crm/forecast-snapshots/backfill/_core.pii.test.ts
@@ -1,41 +1,17 @@
 import { describe, expect, it } from 'vitest';
 
+import { createCrmForecastSnapshotBackfillResult } from '../_core.test-helpers';
+
 import { assertNoCrmForecastSnapshotBackfillPiiKeys } from './_core';
 
 describe('assertNoCrmForecastSnapshotBackfillPiiKeys', () => {
   it('allows aggregate-safe response keys at every depth', () => {
     expect(() =>
-      assertNoCrmForecastSnapshotBackfillPiiKeys({
-        completedAt: '2026-05-15T10:30:01.000Z',
-        dateResults: [
-          {
-            failedWorkItems: 0,
-            snapshotDate: '2026-05-14',
-            snapshotsInserted: 1,
-            status: 'completed',
-            versionConflicts: 0,
-            workItemsConsidered: 1,
-            workItemsDeferred: 0,
-            workItemsSucceeded: 1,
-          },
-        ],
-        datesConsidered: 1,
-        datesDeferred: 0,
-        datesFailed: 0,
-        datesSucceeded: 1,
-        dryRun: false,
-        failedWorkItems: 0,
-        fromDate: '2026-05-14',
-        snapshotsInserted: 1,
-        sourceRunId: 'crm-forecast-snapshot-backfill:tenant-1:2026-05-14:2026-05-14:started',
-        startedAt: '2026-05-15T10:30:00.000Z',
-        tenantId: 'tenant-1',
-        toDate: '2026-05-14',
-        versionConflicts: 0,
-        workItemsConsidered: 1,
-        workItemsDeferred: 0,
-        workItemsSucceeded: 1,
-      })
+      assertNoCrmForecastSnapshotBackfillPiiKeys(
+        createCrmForecastSnapshotBackfillResult({
+          sourceRunId: 'crm-forecast-snapshot-backfill:tenant-1:2026-05-14:2026-05-14:started',
+        })
+      )
     ).not.toThrow();
   });
