+++ apps/web/src/app/api/cron/crm/forecast-snapshots/backfill/_core.test.ts
@@ -1,5 +1,3 @@
-import type { CrmWeightedPipelineRow } from '@interdomestik/domain-crm/reporting';
-import type { CrmForecastSnapshotRow } from '@interdomestik/domain-crm/reporting/types';
 import { describe, expect, it, vi } from 'vitest';
 
 import type { CrmForecastSnapshotWorkItem } from '@/lib/domain-crm/forecast-snapshot-work-items';
@@ -13,84 +11,16 @@ import {
   runCrmForecastSnapshotBackfillCore,
   validateBackfillDateRange,
 } from './_core';
+import {
+  createCrmForecastSnapshotTestDeps,
+  createCrmForecastSnapshotWeightedRow,
+  crmForecastSnapshotTestWorkItem,
+} from '../_core.test-helpers';
 
 const now = new Date('2026-05-15T10:30:00.000Z');
-const workItem: CrmForecastSnapshotWorkItem = {
-  branchId: 'branch-1',
-  currencyCode: 'EUR',
-  pipelineId: 'pipeline-1',
-  tenantId: 'tenant-1',
-};
-
-function weightedRow(overrides: Partial<CrmWeightedPipelineRow> = {}): CrmWeightedPipelineRow {
-  return {
-    agentId: 'agent-1',
-    archivedAt: null,
-    branchId: 'branch-1',
-    currencyCode: 'EUR',
-    currentStageId: 'stage-1',
-    dealId: 'deal-1',
-    forecastCategory: 'commit',
-    isLostStage: false,
-    isWonStage: false,
-    lossReasonId: null,
-    pipelineId: 'pipeline-1',
-    source: null,
-    stageProbability: 50,
-    tenantId: 'tenant-1',
-    utmCampaign: null,
-    utmMedium: null,
-    utmSource: null,
-    valueAmountMinor: 10_000,
-    ...overrides,
-  };
-}
-
-function createDeps(
-  args: {
-    insertResult?: 'success' | 'version_conflict';
-    rows?: CrmWeightedPipelineRow[];
-    workItems?: CrmForecastSnapshotWorkItem[];
-    workItemsDeferred?: number;
-  } = {}
-) {
-  const rows = args.rows ?? [weightedRow()];
-  const workItems = args.workItems ?? [workItem];
-  const workItemRepository = {
-    listWorkItems: vi.fn().mockResolvedValue({
-      workItems,
-      workItemsDeferred: args.workItemsDeferred ?? 0,
-    }),
-  };
-  const reportingRepository = {
-    listAgentLeaderboardRows: vi.fn(),
-    listFunnelConversionRows: vi.fn(),
-    listSourceBreakdownRows: vi.fn(),
-    listStageVelocityRows: vi.fn(),
-    listWeightedPipelineRows: vi.fn().mockResolvedValue(rows),
-    listWinRateRows: vi.fn(),
-  };
-  const snapshotRepository = {
-    insertPipelineSnapshots: vi
-      .fn()
-      .mockImplementation((params: { snapshots: readonly CrmForecastSnapshotRow[] }) => {
-        if (args.insertResult === 'version_conflict') {
-          return Promise.resolve({ success: false, reason: 'version_conflict' });
-        }
-        return Promise.resolve({
-          success: true,
-          snapshots: params.snapshots.map((snapshot, index) => ({
-            ...snapshot,
-            id: `snapshot-${index}`,
-            snapshotVersion: 1,
-          })),
-        });
-      }),
-    listLatestPipelineSnapshots: vi.fn(),
-  };
-
-  return { reportingRepository, snapshotRepository, workItemRepository };
-}
+const workItem: CrmForecastSnapshotWorkItem = crmForecastSnapshotTestWorkItem;
+const weightedRow = createCrmForecastSnapshotWeightedRow;
+const createDeps = createCrmForecastSnapshotTestDeps;
 
 describe('validateBackfillDateRange', () => {
   it('accepts an inclusive historical range only before the current UTC date', () => {
