+++ apps/web/src/app/api/cron/crm/forecast-snapshots/backfill/_core.ts
@@ -19,9 +19,8 @@ import {
 
 import {
   assertNoCrmForecastSnapshotPiiKeys,
-  deriveCrmForecastSnapshotsForWorkItem,
+  buildCrmForecastSnapshotsForWorkItem,
   isCrmForecastSnapshotSoftTimeoutError,
-  loadCrmForecastSnapshotTenantRows,
   throwIfCrmForecastSnapshotAborted,
   withCrmForecastSnapshotTimeout,
   type CrmForecastSnapshotTenantWeightedRows,
@@ -310,31 +309,21 @@ async function processWorkItem(args: {
   tenantRows: CrmForecastSnapshotTenantWeightedRows;
   workItem: CrmForecastSnapshotWorkItem;
 }): Promise<number | 'version_conflict'> {
-  throwIfCrmForecastSnapshotAborted(args.signal, SOFT_TIMEOUT_MESSAGE);
-  const rows = await loadCrmForecastSnapshotTenantRows({
-    reportingInputFor,
-    reportingRepository: args.reportingRepository,
-    snapshotDateEndExclusive: args.snapshotDateEndExclusive,
-    snapshotDateStartInclusive: args.snapshotDateStartInclusive,
-    tenantRows: args.tenantRows,
-    workItem: args.workItem,
-  });
-  throwIfCrmForecastSnapshotAborted(args.signal, SOFT_TIMEOUT_MESSAGE);
-  const input = reportingInputFor(
-    args.workItem.tenantId,
-    args.snapshotDateStartInclusive,
-    args.snapshotDateEndExclusive
-  );
-  const snapshots = deriveCrmForecastSnapshotsForWorkItem({
+  const snapshots = await buildCrmForecastSnapshotsForWorkItem({
+    abortMessage: SOFT_TIMEOUT_MESSAGE,
     createdAt: new Date(args.nowMs()).toISOString(),
     idempotencyKey: buildBackfillSnapshotIdempotencyKey(args.workItem, {
       snapshotDate: args.snapshotDate,
       sourceRunId: args.sourceRunId,
     }),
-    input,
-    rows,
+    reportingInputFor,
+    reportingRepository: args.reportingRepository,
+    signal: args.signal,
     snapshotDate: args.snapshotDate,
+    snapshotDateEndExclusive: args.snapshotDateEndExclusive,
+    snapshotDateStartInclusive: args.snapshotDateStartInclusive,
     sourceRunId: args.sourceRunId,
+    tenantRows: args.tenantRows,
     workItem: args.workItem,
   });
